log-generating-extensions/requirements.txt
numpy
Faker
pytz
tzlocal
kafka-python
fake-factory